requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
streamlit
psycopg2-binary
seaborn